% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talon_plot.R
\name{talon_plot}
\alias{talon_plot}
\title{The 'talon plot': a plot of contributions of survival and recruitment to population growth (lambda)}
\usage{
talon_plot(
  recr_pts,
  surv_pts,
  recr_samples = NULL,
  surv_samples = NULL,
  recr_CI_low,
  recr_CI_high,
  surv_CI_low,
  surv_CI_high,
  col.pts = "black",
  col.samples = alpha("lightgrey", 128),
  plot.samples = TRUE,
  CI_type = c("ellipse", "cross", "none"),
  CI_transform = FALSE,
  col.CI = "grey",
  lim_include_zero_one = TRUE,
  main,
  big_font,
  ylab_offset = if (big_font) 0.25 else 0.15,
  cex = if (big_font) 1.5 else 1,
  cex.pts = 1,
  cex.main = 1.2 * cex,
  cex.lab = cex,
  cex.axis = cex,
  cex.samples = 1,
  font.main = 2,
  mar = if (big_font) c(5.1, 5.1, 5.1, 5.1) else c(4.1, 4.1, 4.1, 4.1),
  ...
)
}
\arguments{
\item{recr_pts, surv_pts}{vector of mean values of recruitment and survival, respectively,
for each data point (data point corresponding e.g. to a temporal occasion)}

\item{recr_samples, surv_samples}{optional. Matrices - rows correspond to MCMC samples,
columns to the same data points as the *_pts variables. Used for plotting and/or calculating 95\% confidence ellipses.}

\item{recr_CI_low, recr_CI_high, surv_CI_low, surv_CI_high}{95\% confidence limits for *_pts variables}

\item{CI_type}{"ellipse", "cross", or "none"}

\item{CI_transform}{only for ellipses now, says whether they should be for bivariate normal
on the scale [recr,surv] for CI_transform = FALSE,
or on the scale [log(recr),logit(surv)] for CI_transform = TRUE}

\item{...}{other graphical parameters}
}
\description{
Makes a so called "talon plot", a plot showing contributions of adult survival and recruitment
to population growth. It is a 2D plot with three axes, population growth being a sum of adult
survival and recruitment. It can reveal important patterns in population dynamics,
see Telenský et al. (2023) <doi:>.
}
\references{
Telenský, T., Storch, D., Klvaňa, P., Reif, J. (2023). Extension of Pradel capture-recapture survival-recruitment model accounting for transients. Methods in Ecology and Evolution. In press.
}
